remove autoClipping folder and merge the autoClipping class into the Clipping class.
