+++ imports/clipping/actions.js
@@ -4,13 +4,16 @@ import { ClippingDB } from '../api/ClippingDB';
 import { mongoUpsert } from '../api/MongoHelper';
 
 const CLIP_CHANGE = 'CLIP_CHANGE';
+const AUTO_CLIP_CHANGE = 'AUTO_CLIP_CHANGE';
 
 export const ActionType = {
   CLIP_CHANGE,
+  AUTO_CLIP_CHANGE,
 };
 
 export function setupClippingDB() {
   api.instance().setupMongoRedux(ClippingDB, CLIP_CHANGE);
+  api.instance().setupMongoRedux(ClippingDB, AUTO_CLIP_CHANGE);
 }
 
 function updateClipping(percentile) {
@@ -27,7 +30,7 @@ function updateClipping(percentile) {
   };
 }
 
-function parseReigsterColormap(resp) {
+function parseReigsterClipping(resp) {
   const { cmd, data } = resp;
   const percentile = data;
   mongoUpsert(ClippingDB, { percentile }, `Resp_${cmd}`);
@@ -41,7 +44,7 @@ function getClipState() {
 
     api.instance().sendCommand(cmd, arg)
       .then((resp) => {
-        parseReigsterColormap(resp);
+        parseReigsterClipping(resp);
       });
   };
 }
@@ -53,10 +56,53 @@ function setupClipping(percentile) {
   };
 }
 
+function updateAutoClipping(autoClip) {
+  return (dispatch, getState) => {
+    // percentile setting belongs to controllerID (controller object) in backend site
+    const controllerID = getState().ImageViewerDB.controllerID;
+    const cmd = `${controllerID}:${Commands.SET_AUTO_CLIP_VALUE}`;
+    const arg = `${Commands.AUTO_CLIP_KEY}:${autoClip}`;
+
+    api.instance().sendCommand(cmd, arg)
+      .then((resp) => {
+        // mongoUpsert(AutoClippingDB, { autoClip }, `Resp_${cmd}`);
+      });
+  };
+}
+
+function parseReigsterAutoClipping(resp) {
+  const { cmd, data } = resp;
+  const autoClip = data;
+  mongoUpsert(ClippingDB, { autoClip }, `Resp_${cmd}`);
+}
+
+function getAutoClipState() {
+  return (dispatch, getState) => {
+    const controllerID = getState().ImageViewerDB.controllerID;
+    const cmd = `${controllerID}:${Commands.GET_AUTO_CLIP_STATE}`;
+    const arg = '';
+
+    api.instance().sendCommand(cmd, arg)
+      .then((resp) => {
+        parseReigsterAutoClipping(resp);
+      });
+  };
+}
+
+function setupAutoClipping(autoClip) {
+  return () => {
+    const cmd = Commands.REGISTER_VIEWER;
+    mongoUpsert(ClippingDB, { autoClip }, `Resp_${cmd}`);
+  };
+}
+
 const actions = {
   updateClipping,
   setupClipping,
   getClipState,
+  updateAutoClipping,
+  setupAutoClipping,
+  getAutoClipState,
 };
 
 export default actions;
